Added soft delete and restore feature
- Users can now restore deleted tickets
- If they want to permanently delete a dicket, they must do it explicitly from the /dashboard/delete page

+++ View.React/ClientApp/src/components/DeleteBlock.js
@@ -1,9 +1,16 @@
 ﻿import axios from 'axios';
 
-const DeleteBlock = ({ ticketId }) => {
+const DeleteBlock = ({ ticketId, removeTicket }) => {
 
     const deleteTicket = async () => {
-        const response = await axios.delete(`https://localhost:5002/api/ticket/${ticketId}`);
+        let response = null;
+        if (!removeTicket) {
+            response = await axios.patch(`https://localhost:5002/api/ticket/${ticketId}`);
+        }
+        else {
+
+            response = await axios.delete(`https://localhost:5002/api/ticket/${ticketId}`);
+        }
         const success = response.status == 204;
         if (success) window.location.reload();
     }
